final exposure calc working
